Project part 3

+++ service-module/src/main/java/com/gmail/yauhenizhukovich/app/service/impl/ReviewServiceImpl.java
@@ -1,6 +1,5 @@
 package com.gmail.yauhenizhukovich.app.service.impl;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import javax.transaction.Transactional;
@@ -8,14 +7,16 @@ import javax.transaction.Transactional;
 import com.gmail.yauhenizhukovich.app.repository.ReviewRepository;
 import com.gmail.yauhenizhukovich.app.repository.model.Review;
 import com.gmail.yauhenizhukovich.app.service.ReviewService;
-import com.gmail.yauhenizhukovich.app.service.model.ReviewDTO;
-import com.gmail.yauhenizhukovich.app.service.util.PaginationUtil;
+import com.gmail.yauhenizhukovich.app.service.model.review.ReviewDTO;
 import com.gmail.yauhenizhukovich.app.service.util.ReviewConversionUtil;
 import org.springframework.stereotype.Service;
 
 import static com.gmail.yauhenizhukovich.app.service.util.PaginationUtil.COUNT_OF_REVIEWS_BY_PAGE;
+import static com.gmail.yauhenizhukovich.app.service.util.PaginationUtil.getCountOfPagesByCountOfObjects;
+import static com.gmail.yauhenizhukovich.app.service.util.PaginationUtil.getStartPositionByPageNumber;
 
 @Service
+@Transactional
 public class ReviewServiceImpl implements ReviewService {
 
     private final ReviewRepository reviewRepository;
@@ -23,9 +24,8 @@ public class ReviewServiceImpl implements ReviewService {
     public ReviewServiceImpl(ReviewRepository reviewRepository) {this.reviewRepository = reviewRepository;}
 
     @Override
-    @Transactional
     public List<ReviewDTO> getReviewsByPage(Integer page) {
-        int startPosition = PaginationUtil.getStartPositionByPageNumber(page, COUNT_OF_REVIEWS_BY_PAGE);
+        int startPosition = getStartPositionByPageNumber(page, COUNT_OF_REVIEWS_BY_PAGE);
         List<Review> reviews = reviewRepository.getObjectsByStartPositionAndMaxResult(startPosition, COUNT_OF_REVIEWS_BY_PAGE);
         return reviews.stream()
                 .map(ReviewConversionUtil::convertDatabaseObjectToDTO)
@@ -33,29 +33,30 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     @Override
-    @Transactional
-    public List<Integer> getPages() {
+    public int getCountOfPages() {
         Long countOfReviews = reviewRepository.getCountOfObjects();
-        return PaginationUtil.getCountOfPages(countOfReviews, COUNT_OF_REVIEWS_BY_PAGE);
+        return getCountOfPagesByCountOfObjects(countOfReviews, COUNT_OF_REVIEWS_BY_PAGE);
     }
 
     @Override
-    @Transactional
-    public void deleteReviewById(Long id) {
+    public boolean deleteReviewById(Long id) {
         Review review = reviewRepository.getById(id);
-        reviewRepository.delete(review);
+        if (review == null) {
+            return false;
+        }
+        return reviewRepository.delete(review);
     }
 
     @Override
-    @Transactional
-    public List<ReviewDTO> updateActivityByIds(Long[] ids) {
-        return Arrays.stream(ids)
+    public List<ReviewDTO> updateStatusByIds(List<Long> ids) {
+        return ids.stream()
                 .map(reviewRepository::getById)
                 .peek(review -> {
-                    boolean actualActive = review.getActive();
-                    review.setActive(!actualActive);
+                    boolean actualStatus = review.getActive();
+                    review.setActive(!actualStatus);
                 })
-                .map(ReviewConversionUtil::convertDatabaseObjectToDTO).collect(Collectors.toList());
+                .map(ReviewConversionUtil::convertDatabaseObjectToDTO)
+                .collect(Collectors.toList());
     }
 
 }
